Updated to v1.3
Updated to v1.3. Changed all SSLv3 to YELLOW because Poodle. Removed the
"Working..." text because it messes with stdout and that makes it more
difficult for other programs to parse the output of this program if you
want to use this program in an automated fashion. Added a NoSchannel
argument, which bypasses the certificate and SChannel stuff and gets
straight to the cipher scanning.